+++ spec/ios_generate_strings_file_from_code_spec.rb
@@ -38,14 +38,14 @@ describe Fastlane::Actions::IosGenerateStringsFileFromCodeAction do
       it 'excludes files matching filters starting with *' do
         test_exclude_patterns(
           filter: ['*.m', '*View.swift'],
-          expected: %w[Sources/AppClass1.swift Pods/SomePod/Sources/PodClass1.swift]
+          expected: %w[Sources/AppClass1.swift Pods/SomePod/Sources/PodClass1.swift Pods/SomePod/Sources/PodLocalizedString.swift]
         )
       end
 
       it 'excludes files matching filters containing * mid-pattern' do
         test_exclude_patterns(
           filter: ['*.m', '*/App*View.swift'],
-          expected: %w[Sources/AppClass1.swift Pods/SomePod/Sources/PodClass1.swift Pods/SomePod/Sources/PodSampleView.swift]
+          expected: %w[Sources/AppClass1.swift Pods/SomePod/Sources/PodClass1.swift Pods/SomePod/Sources/PodLocalizedString.swift Pods/SomePod/Sources/PodSampleView.swift]
         )
       end
     end
@@ -53,36 +53,42 @@ describe Fastlane::Actions::IosGenerateStringsFileFromCodeAction do
 
   context 'when generating .strings files from code' do
     # Helper method for all the test examples in this context group
-    def test_genstrings(params:, expected_dir_name:, expected_logs: nil, expected_failures: nil)
+    #
+    # @param params [Hash] The list of parameters to pass to the call of ios_generate_strings_file_from_code action
+    # @param expected_dir_name [String] The name of the `expected-*` dir in `test-data/…` where the expected output files are located, to compare them with the actual output
+    # @param expected_logs [Array<String>] The list of output lines we expect from the output of `genstrings` and as return value to the action call
+    # @param expected_failure [String] The error message that we expect to be raised by the action, typically via `UI.user_error!`
+    #
+    def test_genstrings(params:, expected_dir_name:, expected_logs: nil, expected_failure: nil)
       # Arrange
       allow_fastlane_action_sh # see spec_helper
       cmd_output = []
       allow(FastlaneCore::UI).to receive(:command_output) { |line| cmd_output << line }
-      user_errors = []
-      allow(FastlaneCore::UI).to receive(:user_error!) { |line| user_errors << line }
+      raised_error_message = nil
 
       Dir.mktmpdir('a8c-wpmrt-ios_generate_strings_file_from_code-') do |tmp_dir|
-        clean_abs_dirs = ->(lines) { lines.map { |l| l.sub(tmp_dir, '<tmpdir>').sub(sample_project_dir, '<testdir>') } }
+        clean_abs_dirs = ->(lines) { Array(lines).map { |l| l.sub(tmp_dir, '<tmpdir>').sub(sample_project_dir, '<testdir>') } }
 
         # Act
         params[:output_dir] = tmp_dir
-        config_items = Fastlane::ConfigurationHelper.parse(described_class, params) # Handles ConfigItem' default values
-        return_value = described_class.run(config_items)
-
-        output_files = Dir[File.join(tmp_dir, '*.strings')]
-        expected_files = Dir[File.join(test_data_dir, expected_dir_name, '*.strings')]
+        return_value = []
+        begin
+          return_value = run_described_fastlane_action(params)
+        rescue FastlaneCore::Interface::FastlaneError => e
+          raised_error_message = e.message
+        end
 
         # Assert: UI.messages, UI.user_error! and return value from the action
+        expect(clean_abs_dirs[raised_error_message]).to eq(Array(expected_failure))
+
         unless expected_logs.nil?
           expect(clean_abs_dirs[cmd_output]).to eq(expected_logs)
           expect(clean_abs_dirs[return_value]).to eq(expected_logs)
         end
 
-        unless expected_failures.nil?
-          expect(clean_abs_dirs[user_errors]).to eq(expected_failures)
-        end
-
         # Assert: same list of generated files
+        output_files = Dir[File.join(tmp_dir, '*.strings')]
+        expected_files = expected_dir_name.nil? ? [] : Dir[File.join(test_data_dir, expected_dir_name, '*.strings')]
         expect(output_files.map { |f| File.basename(f) }.sort).to eq(expected_files.map { |f| File.basename(f) }.sort)
 
         # Assert: each generated file has expected content
@@ -129,8 +135,38 @@ describe Fastlane::Actions::IosGenerateStringsFileFromCodeAction do
     context 'when allowing custom routines' do
       it 'can parse strings from custom routines' do
         test_genstrings(
-          params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, routines: 'PodLocalizedString' },
+          params: {
+            paths: [app_src_dir, pods_src_dir],
+            exclude: ['**/PodLocalizedString.swift'],
+            quiet: true,
+            swiftui: false,
+            routines: 'PodLocalizedString'
+          },
           expected_dir_name: 'expected-custom-routine'
+        )
+      end
+    end
+
+    context 'when requesting custom encoding output' do
+      it 'errors if invalid encoding provided' do
+        test_genstrings(
+          params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, output_encoding: 'unicode' },
+          expected_dir_name: nil,
+          expected_failure: 'unknown encoding name - unicode'
+        )
+      end
+
+      it 'copies the files unchanged if output encoding is already the default UTF-16LE' do
+        test_genstrings(
+          params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, output_encoding: 'utf-16le' },
+          expected_dir_name: 'expected-pods-noswiftui'
+        )
+      end
+
+      it 'convert the files to requested output encoding if not the default UTF-16LE' do
+        test_genstrings(
+          params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, output_encoding: 'utf-8' },
+          expected_dir_name: 'expected-utf8-encoding'
         )
       end
     end
@@ -163,26 +199,22 @@ describe Fastlane::Actions::IosGenerateStringsFileFromCodeAction do
       it 'does not fail if any error is in the output when `fail_on_error` is off' do
         expected_logs = [
           %(Key "app.key5" used with multiple values. Value "app value 5\\nwith multiple lines." kept. Value "app value 5\\nwith multiple lines, and different value than in Swift" ignored.),
-          %(genstrings: error: bad entry in file <testdir>/Pods/SomePod/Sources/PodClass1.swift (line = 3): Argument is not a literal string.),
+          %(genstrings: error: bad entry in file <testdir>/Pods/SomePod/Sources/PodLocalizedString.swift (line = 3): Argument is not a literal string.),
         ]
         test_genstrings(
           params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, routines: 'PodLocalizedString', fail_on_error: false },
           expected_dir_name: 'expected-custom-routine',
           expected_logs: expected_logs,
-          expected_failures: []
+          expected_failure: nil
         )
       end
 
       it 'fails if there is any error in the output in `fail_on_error` mode, even in quiet mode' do
-        expected_logs = [
-          %(Key "app.key5" used with multiple values. Value "app value 5\\nwith multiple lines." kept. Value "app value 5\\nwith multiple lines, and different value than in Swift" ignored.),
-          %(genstrings: error: bad entry in file <testdir>/Pods/SomePod/Sources/PodClass1.swift (line = 3): Argument is not a literal string.),
-        ]
+        expected_failure = %(genstrings: error: bad entry in file <testdir>/Pods/SomePod/Sources/PodLocalizedString.swift (line = 3): Argument is not a literal string.)
         test_genstrings(
           params: { paths: [app_src_dir, pods_src_dir], quiet: true, swiftui: false, routines: 'PodLocalizedString', fail_on_error: true },
-          expected_dir_name: 'expected-custom-routine',
-          expected_logs: expected_logs,
-          expected_failures: [expected_logs.last]
+          expected_dir_name: nil,
+          expected_failure: expected_failure
         )
       end
 
@@ -196,7 +228,7 @@ describe Fastlane::Actions::IosGenerateStringsFileFromCodeAction do
           params: { paths: [app_src_dir, pods_src_dir], quiet: false, swiftui: false, fail_on_error: true },
           expected_dir_name: 'expected-pods-noswiftui',
           expected_logs: expected_logs,
-          expected_failures: []
+          expected_failure: nil
         )
       end
     end