+++ spec/test-data/translations/ios_generate_strings_file_from_code/sample-project/Pods/SomePod/Sources/PodLocalizedString.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+func PodLocalizedString(_ key: String, value: String? = nil, comment: StaticString) -> String {
+    Bundle(for: BundleToken.self).localizedString(forKey: key, value: value, table: nil)
+}